Version 0.14.0
New version 0.14.0 with several changes:
Compiler now accepts the older (* *) style comments in addition to { } and single-line // comments. If you fail to close a comment the compiler not only reports this at the end of file, but the line and position where the last comment was open. If you accidentally use a } without an open brace the compiler says so.
Added local time function to Sysutils, allowing the compiler to report when the compiler was started and when it ended
Many new statistics reported: Compiler prints number of lines for each file and total number of lines compiled. It also reports compilation speed in lines per second, which on my several year old computer was more than 48,000 lines/second. The compiler can compile itself - a 12,000 line program in about six modules - in 0.25 seconds. This release also includes a pre-compiled compiler so you don't have to have Delphi or Free Pascal to use it.
Code sizes are reported in decimal and hexadecimal.